refactor: move aligend_layer_service_manager to its own .ex
